# perf(core-database-postgres): faster count method