--- a/packages/core-database-postgres/src/repositories/blocks.ts
+++ b/packages/core-database-postgres/src/repositories/blocks.ts
@@ -8,7 +8,7 @@ export class BlocksRepository extends Repository implements Database.IBlocksRepo
     public async search(params: Database.ISearchParameters): Promise<{ rows: Interfaces.IBlockData[]; count: number }> {
         // TODO: we're selecting all the columns right now. Add support for choosing specific columns, when it proves useful.
         const selectQuery = this.query.select().from(this.query);
-        const selectQueryCount = this.query.select(this.query.count().as("cnt")).from(this.query);
+        const selectQueryCount = this.query.select(this.query.id).from(this.query);
         // Blocks repo atm, doesn't search using any custom parameters
         const parameterList = params.parameters.filter(o => o.operator !== Database.SearchOperator.OP_CUSTOM);
 
@@ -20,11 +20,11 @@ export class BlocksRepository extends Repository implements Database.IBlocksRepo
             } while (!first.operator && parameterList.length);
 
             if (first) {
-                for (const query of [ selectQuery, selectQueryCount ]) {
+                for (const query of [selectQuery, selectQueryCount]) {
                     query.where(this.query[this.propToColumnName(first.field)][first.operator](first.value));
                 }
                 for (const param of parameterList) {
-                    for (const query of [ selectQuery, selectQueryCount ]) {
+                    for (const query of [selectQuery, selectQueryCount]) {
                         query.and(this.query[this.propToColumnName(param.field)][param.operator](param.value));
                     }
                 }