@vasild 2019-07-18
I am afraid that there is no silver bullet for the problem with the slow `COUNT(*)`.

`SELECT COUNT(*) FROM (SELECT * FROM table);` is not faster than `SELECT COUNT(*) FROM table;`. The former is either slower, or with a smart enough optimizer, equivalent to the later.